+++ x/ccv/consumer/keeper/migration.go
@@ -21,23 +21,27 @@ func NewMigrator(ccvConsumerKeeper Keeper, ccvConsumerParamSpace paramtypes.Subs
 	return Migrator{ccvConsumerKeeper: ccvConsumerKeeper, ccvConsumerParamSpace: ccvConsumerParamSpace}
 }
 
+// Migrate1to2 migrates x/ccvconsumer state from consensus version 1 to 2.
+func (m Migrator) Migrate1to2(ctx sdk.Context) error {
+	return m.ccvConsumerKeeper.MigrateConsumerPacketData(ctx)
+}
+
 // MigrateConsumerPacketData migrates consumer packet data according to
 // https://github.com/cosmos/interchain-security/pull/1037
 //
 // Note an equivalent migration is not required for providers.
-func (k Keeper) MigrateConsumerPacketData(ctx sdk.Context) {
+func (k Keeper) MigrateConsumerPacketData(ctx sdk.Context) error {
 	// deserialize packet data from old format
 	var depreciatedType ccvtypes.ConsumerPacketDataList
 	store := ctx.KVStore(k.storeKey)
 	bz := store.Get([]byte{consumertypes.PendingDataPacketsBytePrefix})
 	if bz == nil {
 		ctx.Logger().Info("no pending data packets to migrate")
-		return
+		return nil
 	}
 	err := depreciatedType.Unmarshal(bz)
 	if err != nil {
-		// An error here would indicate something is very wrong
-		panic(fmt.Errorf("failed to unmarshal pending data packets: %w", err))
+		return fmt.Errorf("failed to unmarshal pending data packets: %w", err)
 	}
 
 	// Delete old data
@@ -48,6 +52,7 @@ func (k Keeper) MigrateConsumerPacketData(ctx sdk.Context) {
 	for _, data := range depreciatedType.List {
 		k.AppendPendingPacket(ctx, data.Type, data.Data)
 	}
+	return nil
 }
 
 // TODO: the following hackyness could be removed if we're able to reference older versions of ICS.